backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pydantic[email]==2.10.4
email-validator==2.2.0
pydantic-settings==2.7.0
celery==5.4.0
redis==5.2.1
pandas==2.2.3
numpy==2.2.1
requests==2.32.3
python-dotenv==1.0.1
reportlab==4.2.5
matplotlib==3.10.0
alpha-vantage==2.3.1
